expenses-bot/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
deep-translator==1.11.4
et_xmlfile==2.0.0
idna==3.10
numpy==2.2.0
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
tqdm==4.67.1
tzdata==2024.2
urllib3==2.2.3
